amf/requirements.txt
# Core ML runtime (choose variant by OS)
# Linux/Windows
tensorflow>=2.16,<3; platform_system != "Darwin"
# macOS (Apple Silicon or Intel)
tensorflow-macos>=2.16,<3; platform_system == "Darwin"
# Metal acceleration for macOS
tensorflow-metal>=1.1; platform_system == "Darwin"

# Arrays, science stack
numpy>=1.26,<3
scipy>=1.10
h5py>=3.8,<4

# Data & plots
pandas>=2.0
matplotlib>=3.7
Pillow>=10

# Image/IO
opencv-python>=4.8
pyvips>=2.2
PyYAML>=6.0
psutil>=5.9
imagesize>=1.4

# ML utils
scikit-learn>=1.1

# Visualization of model graph (optional, requires system Graphviz)
pydot>=1.4
graphviz>=0.20

# TensorBoard (optional)
tensorboard>=2.16,<2.17

# NOTE:
# - Removed legacy standalone keras, tf-agents, tensorflow-probability, gym, dm-tree,
#   gin-config, and google-* packages. The project now uses tf.keras exclusively.
# - On macOS, ensure you have system dependencies:
#   brew install vips graphviz  # for pyvips and plot_model
